+++ src/main.c
@@ -23,56 +23,9 @@
 #include "web/settings_api.h"
 #include "settings/settings.h"
 #include "settings/devices.h"
+#include "web/web_api.h"
 
 static const char * TAG = "MAIN";
-
-esp_err_t init_routes(httpd_handle_t server, rest_server_context_t *rest_context)
-{
-    httpd_uri_t common_get_uri = {
-        .uri = "/",
-        .method = HTTP_GET,
-        .handler = rest_common_get_handler,
-        .user_ctx = rest_context
-    };
-
-    httpd_uri_t get_css_uri = {
-        .uri = "/modest.css",
-        .method = HTTP_GET,
-        .handler = rest_common_get_handler,
-        .user_ctx = rest_context
-    };
-
-    httpd_uri_t get_js_uri = {
-        .uri = "/main.js",
-        .method = HTTP_GET,
-        .handler = rest_common_get_handler,
-        .user_ctx = rest_context
-    };
-
-    httpd_uri_t get_settings_uri = {
-        .uri = "/settings.html",
-        .method = HTTP_GET,
-        .handler = rest_common_get_handler,
-        .user_ctx = rest_context
-    };
-
-    httpd_uri_t get_readings_uri = {
-        .uri = "/readings.html",
-        .method = HTTP_GET,
-        .handler = rest_common_get_handler,
-        .user_ctx = rest_context
-    };
-
-    httpd_register_uri_handler(server, &common_get_uri);
-    httpd_register_uri_handler(server, &get_css_uri);
-    httpd_register_uri_handler(server, &get_settings_uri);
-    httpd_register_uri_handler(server, &get_js_uri);
-    httpd_register_uri_handler(server, &get_readings_uri);
-
-    init_settings_routes(server, rest_context);
-    return ESP_OK;
-}
-
 
 /***************************
  * MAIN
@@ -119,22 +72,6 @@ void app_main()
 #endif
     while(1)
     {
-        /*
-        uint8_t data[128];
-        int len = rs485_read(data);
-        protocol_data_raw_t data_parsed;
-        if(protocol_message_parse((char*)data, &data_parsed))
-        {
-            printf("A mensagem %s é para mim! \n", (char *)data);
-        }
-
-        data[len] = '\0';
-        if(len > 0)
-        {
-            printf("Recebido: %s\n", (char*)data);
-        }
-        printf("pulses: %d - litros: %f\n", flowsensor_get_pulses(), flowsensor_get_litros());
-        */
         vTaskDelay(pdMS_TO_TICKS(1000));
     }
 }